workspaces/FFS1_ColliderDataSystems/FFS2_ColliderBackends_MultiAgentChromeExtension/ColliderDataServer/pyproject.toml
[project]
name = "collider-data-server"
version = "0.1.0"
description = "Collider Data Server — REST + SSE API for the Collider ecosystem"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "aiosqlite>=0.21.0",
    "alembic>=1.15.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.9.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.0.0",
    "pyyaml>=6.0.0",
    "bcrypt>=4.0.0",
    "PyJWT>=2.8.0",
    "sqlite-web>=0.6.3",
    "grpcio>=1.70.0",
    "grpcio-tools>=1.70.0",
]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
